backend/discord-bot/requirements.txt
PyNaCl>=1.5.0
psycopg2-binary